pyproject.toml
[project]
name = "fmov"
version = "0.2.0"
description = "A simple, optimized way to programmatically create video with python. Includes interactive preview and animation debugging tools."
authors = [
    {name = "Dylan DiBeneditto", email = "dibenedittod@gmail.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pillow",
    "ffmpeg-python",
    "numpy",
    "PyQt5",
    "rich"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.setuptools]
packages = {find = {where = ["src"]}}
